requirements.txt
peft==0.15.2
transformers==4.52.3
python-dotenv
b2sdk
